// pab_ng-main/pab/src/app/shared/enum/kostenart.enum.ts
export enum KostenartEnum {
  PROJEKTZEITEN = 'Projektzeiten',
  REISEZEITEN = 'Reisezeiten',
  REISEKOSTEN = 'Reisekosten',
  SONDERARBEITSZEITEN = 'Sonderarbeitszeiten',
  RUFBEREITSCHAFT = 'Rufbereitschaften',
  SONSTIGE = 'Sonstige',
  SKONTO = 'Skonto',
  RABATTE = 'Rabatte',
  FAKTFAEHIGELEISTUNGEN = 'FaktfaehigeLeistung',
}